+++ provision/env.sh
@@ -1,11 +1,31 @@
+#!/usr/bin/env bash
 # Environment-specific settings used by the provisioning process
 
-## EXAMPLE FILE ONLY ##
-
+#
 # Required
-DB_PASS='password' # do not use this password
+#
 
+PUBLIC_KEY=''
+
+#
 # Optional
-PUBLIC_KEY='ssh-rsa AAAAB3NzaC1yc2EAAAABJQAAAQBjSyMHVRz2JJl7QalvNbaQvnqL8T5VZoBtPqQqMftWLbtwF7AAUhlWM/x1rctRJbW6bLQQxFVMCqs5ovZITiVtR4BXr8+6v8V8xPh68nA6oIjCL7w14v6DxHEAOeCsK7PiwSwW9I8jwbrzRLmNj1yGLvAwSjgV5vVjou34Qijax/R/93Koj6+vHJcQ5OPH3S5y2OJZVYILNGbLIgLrTTfghdUHZSWJ2AbKIuYd3XsaRtlgsJgHPQ8TgtB95cf9ocDl0EXaKFsNq+jmLTYKw6ryamE0ugE+d75tJDXnfOiG72oFgYHk/84ENK3vP3Fq4m73RNoSIs/c8gef9qpcbt1d alex@church.id.au'
-TIMEZONE='Australia/Sydney'
+#
+
 DEBUG=1
+TIME_ZONE='Australia/Sydney'
+
+# A secret key will be generated automatically if not provided.
+#SECRET_KEY=''
+
+# A database password will be generated automatically if not provided.
+#DB_PASS=''
+
+# The template file for the env.py file written during provisioning, relative
+# to provision/templates/.
+# Defaults env.py.txt
+#ENV_PY_TEMPLATE=''
+
+#
+# Non-customisable settings will be created below here by
+# the provisioning process
+#