+++ provision/scripts/bootstrap.sh
@@ -3,145 +3,118 @@
 # vagrant-django
 # Vagrant provisioning for Django projects.
 # https://github.com/oogles/vagrant-django
-# v0.3.2
+# v0.4.0
 
-PROJECT_NAME="$1"
-BUILD_MODE="$2"
+# Ensure all provisioning scripts are executable
+chmod 774 -R /opt/app/src/provision/scripts/
 
-if [[ ! "$PROJECT_NAME" ]]; then
-    echo "--------------------------------------------------"
-    echo "ERROR: No project name provided."
-    echo "--------------------------------------------------"
-    exit 1
-fi
-
-if [[ ! "$BUILD_MODE" ]]; then
-    echo "--------------------------------------------------"
-    echo "ERROR: No build mode provided."
-    echo "--------------------------------------------------"
-    exit 1
-elif [[ "$BUILD_MODE" != "project" && "$BUILD_MODE" != "app" ]]; then
-    echo "--------------------------------------------------"
-    echo "ERROR: Unknown build mode \"$BUILD_MODE\"."
-    echo "--------------------------------------------------"
-    exit 1
-fi
-
-# Get environment-specific variables from config.
-# For a full description of the available variables, and the effects they have
-# on the provisioning process, see the docs at https://vagrant-django.readthedocs.org/.
-# DB_PASS:         Required. The password to use for the app database created in 
-#                  PostgreSQL.
-# PUBLIC_KEY:      Optional. A custom public key to install in .ssh/authorized_keys.
-# DEBUG:           Optional. Set to 1 to set 'DEBUG': True in the environment-specific
-#                  settings file for the project.
-# TIME_ZONE:       Optional. The server time zone. Defaults to Australia/Sydney.
-if [[ -f /vagrant/provision/env.sh ]]; then
-    source /vagrant/provision/env.sh
-else
-    echo "--------------------------------------------------"
-    echo "ERROR: Missing required environment-specific config file provision/env.sh."
-    echo "--------------------------------------------------"
-    exit 1
-fi
-
-if [[ ! "$DB_PASS" ]]; then
-    echo "--------------------------------------------------"
-    echo "ERROR: No DB_PASS variable defined in provision/env.sh."
-    echo "--------------------------------------------------"
-    exit 1
-fi
-
-if [[ "$BUILD_MODE" == "app" ]]; then
-    DEBUG=1
-elif [[ "$DEBUG" && "$DEBUG" -eq 1 ]]; then
-    DEBUG=1
-else
-    DEBUG=0
-fi
+# Define a function for executing provisioning scripts. The first argument
+# should be the path to the script file to execute. Remaining arguments will be
+# passed through to the script. The primary reason for encapsulating these calls
+# in a function is to exit this outer script if one of the executed scripts
+# returns a non-zero exit code.
+function run_script() {
+    "$1" "${@:2}"
+    if [[ $? != 0 ]]; then
+        echo " "
+        echo "=================================================="
+        exit 1
+    fi
+}
 
 echo " "
 echo "=================================================="
 echo " "
-echo "START PROVISION FOR \"$PROJECT_NAME\""
-echo " "
+echo "START PROVISION"
 
-echo " --- Adding SSH public key ---"
-if [[ "$PUBLIC_KEY" ]]; then
-    if ! grep -Fxq "$PUBLIC_KEY" .ssh/authorized_keys ; then
-        echo "$PUBLIC_KEY" >> .ssh/authorized_keys
-        echo "Done"
-    else
-        echo "Public key already present in authorized_keys."
-    fi
-else
-    echo "No PUBLIC_KEY env setting defined. Custom public key not added."
-fi
+# Define common settings, passing the arguments that were passed to this script
+run_script /opt/app/src/provision/scripts/settings.sh "$1"
+
+# Source the defined settings
+source /tmp/env.sh
+
+# Add/update apt repos - get software current before doing anything
+run_script "$PROVISION_DIR/scripts/apt.sh"
+
+# Setup the "webmaster" user and home directory
+run_script "$PROVISION_DIR/scripts/user.sh"
+
+# Configure SSH
+run_script "$PROVISION_DIR/scripts/ssh.sh" "$PUBLIC_KEY"
 
 echo " "
-echo " --- Setting time zone ---"
+echo " --- Set time zone ---"
 if [[ ! "$TIME_ZONE" ]]; then
     TIME_ZONE='Australia/Sydney'
 fi
-echo "$TIME_ZONE" | tee /etc/timezone && dpkg-reconfigure --frontend noninteractive tzdata
+echo "$TIME_ZONE" | tee /etc/timezone
+dpkg-reconfigure --frontend noninteractive tzdata
 
-# Add/update apt repos
-/vagrant/provision/scripts/apt.sh
+# Set up app directory
+run_script "$PROVISION_DIR/scripts/app-dir.sh"
+
+# Write environment settings file
+run_script "$PROVISION_DIR/scripts/write-env-settings.sh" "$SECRET_KEY" "$TIME_ZONE" "$DB_PASS"
+
+# Enable a firewall in production environments
+if [[ "$DEBUG" -eq 0 ]]; then
+    run_script "$PROVISION_DIR/scripts/firewall.sh"
+fi
 
 # Some basic installs
-/vagrant/provision/scripts/install.sh
+run_script "$PROVISION_DIR/scripts/install.sh"
+
+# Install and configure supervisor
+run_script "$PROVISION_DIR/scripts/supervisor.sh"
 
 # Install and configure database
-/vagrant/provision/scripts/database.sh "$PROJECT_NAME" "$DB_PASS"
+run_script "$PROVISION_DIR/scripts/database.sh" "$DB_PASS"
 
-# If a project-specific provisioning file is present, run it
-if [[ -f /vagrant/provision/project.sh ]]; then
+# If a project-specific provisioning file is present, ensure it is executable
+# and run it
+if [[ -f "$PROVISION_DIR/project.sh" ]]; then
     echo " "
-    echo " --- Running project-specific configuration ---"
-	/vagrant/provision/project.sh "$PROJECT_NAME" "$BUILD_MODE" "$DEBUG"
+    echo " --- Run project-specific configuration ---"
+    chmod 774 "$PROVISION_DIR/project.sh"
+    run_script "$PROVISION_DIR/project.sh"
 fi
 
 # Install and configure virtualenv and install python dependencies.
 # Must run after postgres is installed if installing psycopg2, and after image
 # libraries if installing Pillow.
-/vagrant/provision/scripts/pip-virtualenv.sh "$PROJECT_NAME" "$BUILD_MODE" "$DEBUG"
+run_script "$PROVISION_DIR/scripts/pip-virtualenv.sh"
 
 # Install and configure nodejs/npm and install node dependencies, if the project
 # makes use of them
-if [[ -f /vagrant/package.json ]]; then
-    /vagrant/provision/scripts/node-npm.sh "$DEBUG"
+if [[ -f "$SRC_DIR/package.json" ]]; then
+    run_script "$PROVISION_DIR/scripts/node-npm.sh"
 fi
 
-# Copy necessary config files
-if [[ -d /vagrant/provision/conf ]]; then
-    /vagrant/provision/scripts/copy-conf.sh
+# Install nginx and gunicorn for production environments. Must run after
+# virtualenv is installed.
+if [[ "$DEBUG" -eq 0 ]]; then
+    run_script "$PROVISION_DIR/scripts/nginx-gunicorn.sh"
 fi
 
-# Write environment settings file
-if [[ "$BUILD_MODE" == "project" ]]; then
-    /vagrant/provision/scripts/write-env-settings.sh "$PROJECT_NAME" "$DEBUG" "$DB_PASS" "$TIME_ZONE"
-fi
-
+# Start supervisor now that program files are in place and any necessary
+# installations/configurations for those programs are complete
 echo " "
-echo " --- Running migrations ---"
-if [[ -f "/vagrant/manage.py" ]]; then
-    su - vagrant -c "source ~/.virtualenvs/$PROJECT_NAME/bin/activate && /vagrant/manage.py migrate"
-else
-    echo "--------------------------------------------------"
-    echo "WARNING: No manage.py file detected."
-    echo "No migrations have been run."
-    echo "--------------------------------------------------"
-fi
-
-# Add custom scripts
-if [[ ! -d bin ]] ; then
-    su - vagrant -c "mkdir ~/bin"
-fi
-
-/vagrant/provision/scripts/bin/shell+.sh
-/vagrant/provision/scripts/bin/runserver+.sh
+echo " --- Start supervisor ---"
+service supervisor start
+echo "Done"
 
 echo " "
 echo "END PROVISION"
 echo " "
 echo "=================================================="
+
+if [[ "$DEBUG" -eq 0 ]]; then
+    echo " "
+    echo "IMPORTANT"
+    echo "The \"webmaster\" user account requires a password for sudo access."
+    echo "NOTE: It does NOT require one for SSH access (password-based SSH access is disabled)."
+    echo "Set a password using the following command:"
+    echo "passwd"
+    echo " "
+    echo "=================================================="
+fi