+++ provision/scripts/node-npm.sh
@@ -1,9 +1,10 @@
 #!/usr/bin/env bash
 
+# Source global provisioning settings
+source /tmp/env.sh
+
 echo " "
 echo " --- Install node.js/npm ---"
-
-DEBUG="$1"
 
 # Install node and update npm
 apt-get -qq install nodejs
@@ -13,12 +14,13 @@ npm install npm -g --quiet
 # This avoids multiple issues when using a Windows host.
 # NOTE: The name of the linked directory seems to want to be "node_modules".
 # Anything else causes issues.
-NODE_MODULES_PATH="/home/vagrant/node_modules"
-NODE_LINK_PATH="/vagrant/node_modules"
+NODE_MODULES_PATH="$APP_DIR/node_modules"
+NODE_LINK_PATH="$SRC_DIR/node_modules"
 
 if [[ ! -d "$NODE_MODULES_PATH" ]]; then
     mkdir "$NODE_MODULES_PATH"
-    chown vagrant:vagrant "$NODE_MODULES_PATH"
+    chown www-data:www-data "$NODE_MODULES_PATH"
+    chmod g+w "$NODE_MODULES_PATH"
 fi
 
 # Create a symlink to the relocated node_modules directory.
@@ -30,11 +32,11 @@ fi
 
 ln -s "$NODE_MODULES_PATH" "$NODE_LINK_PATH"
 
-# Install project dependencies
+## Install project dependencies
 echo " "
 echo " --- Install node.js dependencies ---"
 if [[ "$DEBUG" -eq 1 ]]; then
-    su - vagrant -c "cd /vagrant/ && npm install --quiet"
+    su - webmaster -c "cd $SRC_DIR && npm install --quiet"
 else
-    su - vagrant -c "cd /vagrant/ && npm install --production --quiet"
+    su - webmaster -c "cd $SRC_DIR && npm install --production --quiet"
 fi